Added my selenium test file for movement
As mentioned in Discord, I (dummy) made a seperate file for the movement testing so I am just adding it as a seperate file so that I do not have to replicate the tests. This will be the only time in which my tests will be on a seperate file.