badmax implemented
The user story for bad max (sh-14) is fully impelented and seems to be working. Now just need to run some tests on it. Also, I want to add functionality that the user can scan for him and see if he is near their position represented by an image on the map?